requirements.txt
reflex>=0.6.6,<0.9
requests>=2.31.0
pyautogui>=0.9.54
pynput>=1.7.6
Pillow>=9.5.0
pyperclip>=1.8.2
opencv-python>=4.9.0.80
Flask>=3.0.0
pywebview>=5.0
